// WedUI/src/app/patient/Models/IPatient.ts
export interface IPatient {
    id: string;
    name: string;
    fileNo: number;
    citizenId: string;
    birthDate: string;
    gender: number;
    nationality: string;
    phoneNumber: string;
    email: string;
    country: string;
    city: string;
    street: string;
    address1: string;
    address2: string;
    contactPerson: string;
    contactRelation: string;
    contactPhone: string;
    firstVisitDate: string;
    recordCreationDate: string;
}